feat(v3): inference cache fix + SFT data pipeline + eval/sweep tooling
模型 / 配置:
- modeling: SNNBlock conv_state ring buffer + SNNAttentionDecoderLayer
  pos_offset 支持 stateful generation; NeuronSparkForCausalLM.generate_cached
  + per-tensor mixed precision from_pretrained
- config: num_hidden_layers + use_cache=False (HF GenerationMixin 不建 cache)

SFT 数据:
- pretokenize_sft_v3.py: parquet/arrow → uint32 .bin + uint8 .mask.bin + uint64 .bos.idx
- build_v3_sft_mix.py: 多源混合 (R1 EN/ZH thinking + smoltalk2 Qwen3 + WildChat GPT-4)
- filter_to_2048.py: 按 [min, cap] 严格过滤无截断
- nsdata/BinnedSFTDataset: memmap 读三件套, 与 SFTDataset 同接口
- train_sft.py: auto-detect binned vs runtime tokenize

工具:
- eval_queue.py: 13-task lm-eval 串行队列
- sweep_inference.py + compare_sweep.py: 多温度多超参 sweep + cached/nocache 对比
- test_inference_cache.py: cache 正确性 sanity check (greedy bit-level)
- download_all/remaining_benchmarks.py: benchmark 数据下载

文档: docs/INFERENCE_WORKFLOW.md (H200→本地→HF 工作流)

+++ generate.py
@@ -29,7 +29,7 @@ def load(ckpt_path: str, tokenizer_path: str | None = None, device: str = "cuda"
     tok_src = tokenizer_path or ckpt_path
     tokenizer = AutoTokenizer.from_pretrained(tok_src, trust_remote_code=True)
     model = AutoModelForCausalLM.from_pretrained(
-        ckpt_path, dtype=torch.float32, trust_remote_code=True,
+        ckpt_path, trust_remote_code=True,
     )
     model = model.to(device).eval()
     return model, tokenizer
